requirements.txt
faster-whisper
groq
python-dotenv
moviepy
textwrap
pillow
streamlit